added several new widgetastic controls
