+++ widgetastic_manageiq.py
@@ -20,8 +20,7 @@ from widgetastic.widget import (
 from widgetastic.utils import ParametrizedLocator
 from widgetastic.xpath import quote
 from widgetastic_patternfly import (
-    Accordion as PFAccordion, CandidateNotFound, BootstrapTreeview, Button, Input, BootstrapSelect,
-    Dropdown)
+    Accordion as PFAccordion, CandidateNotFound, BootstrapTreeview, Button, Input, BootstrapSelect)
 from cached_property import cached_property
 from utils.log import logger
 
@@ -881,141 +880,90 @@ class PaginationPane(View):
         return self.paginator.page_info()[1]
 
 
-class ToolBarViewSelector(View):
+class Stepper(Widget):
+    """ A CFME Stepper Control
+
+        stepper = Stepper(locator='//')
     """
-    represents toolbar's view selector control.
-    """
-    list_view = Button(title='List View')
-    grid_view = Button(title='Grid View')
-    tile_view = Button(title='Tile View')
+    def __init__(self, parent, locator, logger=None):
+        super(Stepper, self).__init__(parent=parent, logger=logger)
+
+        self._stepper_control = self.browser.element(locator, parent=parent)
+
+        self.minus_button = Button(parent=self._stepper_control, text='-')
+        self.plus_button = Button(parent=self._stepper_control, text='+')
+        self.value_field = Input(parent=self._stepper_control, locator='.//input[contains(@class, '
+                                                                       '"bootstrap-touchspin"]')
 
     def __locator__(self):
-        return './/div[contains(@class, "toolbar-pf-view-selector")]'
+        return self._stepper_control
 
-    def select(self, item):
-        if item == self.list_view.title:
-            self.list_view.click()
-        elif item == 'Grid View':
-            self.grid_view.click()
-        elif item == 'Tile View':
-            self.grid_view.click()
+    @property
+    def read(self):
+        return int(self.value_field.read())
+
+    def decrease(self):
+        self.minus_button.click()
+
+    def increase(self):
+        self.plus_button.click()
+
+    def set_value(self, value):
+        steps = int(value) - self.read
+        if steps == 0:
+            return
+        elif steps > 0:
+            operation = self.increase
         else:
-            raise ValueError('Incorrect value passed')
+            operation = self.decrease
+
+        steps = abs(steps)
+        for step in range(steps):
+            operation()
+
+
+class RadioButton(Button):
+    """CFME Radio Button Control
+
+    """
+    def __locator__(self):
+        return (
+            './/*[self::input and @type="radio" '
+            'and ({})]'.format(self.locator_conditions))
+
+
+class RadioGroup(Widget):
+    """ CFME Radio Group Control
+
+    """
+    def __init__(self, parent, logger=None):
+        super(RadioGroup, self).__init__(parent=parent, logger=logger)
+
+    @property
+    def buttons(self):
+        radio_buttons = []
+        for btn in self.browser.elements('.//input[@type="radio"]', parent=self.parent):
+            radio_buttons.append(RadioButton(text=btn.text))
+        return radio_buttons
+
+    @property
+    def button_names(self):
+        return [self.browser.text(btn) for btn in self.buttons]
 
     @property
     def selected(self):
-        """
-        goes thru buttons and returns the title of active button
-        Returns: currently selected view
+        btn = [btn for btn in self.buttons if 'ng-valid-parse' in self.browser.classes(btn)][-1]
+        return self.browser.text(btn)
 
-        """
-        return [btn.title for btn in (self.list_view, self.grid_view, self.tile_view)
-                if btn.active][-1]
-
-
-class ProviderToolBar(View):
-    """
-    represents provider toolbar and its controls
-    """
-    configuration = Dropdown(text='Configuration')
-    policy = Dropdown(text='Policy')
-    authentication = Dropdown(text='Authentication')
-    download = Dropdown(text='Download')
-
-    @View.nested
-    class view_selector(ToolBarViewSelector):  # NOQA
-        pass
-
-
-class DetailsProviderToolBar(View):
-    """
-    represents provider toolbar and its controls
-    """
-    monitoring = Dropdown(text='Monitoring')
-    configuration = Dropdown(text='Configuration')
-    reload = Button(title='Reload Current Display')
-    policy = Dropdown(text='Policy')
-    authentication = Dropdown(text='Authentication')
-
-    @View.nested
-    class view_selector(ToolBarViewSelector):  #NOQA
-        # todo: there should be another ViewSelector. to add it later
-        pass
-
-
-class Search(Widget):
-    # todo: add it later
-    pass
-
-
-class Items(View):
-    """
-    should represent the view with different items like providers
-    """
-    @View.nested
-    class search(Search):  # NOQA
-        pass
-
-
-class BaseSideBar(View):
-    """
-    represents left side bar. it usually contains navigation, filters, etc
-    """
-    pass
+    def select(self, button_name):
+        if button_name != self.selected:
+            btn = [btn for btn in self.buttons if self.browser.text(btn) == button_name][-1]
+            btn.click()
 
 
 class BreadCrumb(Widget):
+    """ CFME BreadCrumb navigation control
+
     """
-    represents breadcrumb navigation control
-    """
+
     pass
-
-
-class TimelinesFilter(View):
-
-    def __init__(self, parent, filter_type, logger=None):
-        super(TimelinesFilter, self).__init__(parent=parent, logger=logger)
-        if filter_type in ('Management Events', 'Policy Events'):
-            self._filter_type = filter_type
-        else:
-            raise ValueError('incorrect filter type is passed')
-
-    # common
-    event_type = BootstrapSelect(id='tl_show')
-    event_category = BootstrapSelect(id='tl_category_management')
-    time_period = Stepper()
-    time_range = BootstrapSelect(id='tl_range')
-    time_position = BootstrapSelect(id='tl_timepivot')
-    date_picker = DatePicker()
-    apply = Button("Apply")
-    # management controls
-    detailed_events = Checkbox(name='showDetailedEvents')
-    # policy controls
-    event_category = BootstrapSelect(id='tl_category_policy')
-    event_status = RadioGroup()
-
-
-class TimelinesChart(View):
-    # todo: to add widgets for all controls
-    # currently only events collection is available
-    pass
-
-
-class Timelines(View):
-    """
-    represents Timelines page
-    """
-    @View.nested
-    class sidebar(BaseSideBar):  # NOQA
-        pass
-
-    breadcrumb = BreadCrumb()
-
-    @View.nested
-    class filter(TimelinesFilter):  # NOQA
-        pass
-
-    @View.nested
-    class chart(TimelinesChart):  # NOQA
-        pass
-
